backend/requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==4.2.3
django-password-validators==1.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
gunicorn==21.2.0
isort==5.12.0
mccabe==0.7.0
packaging==23.1
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
